Update for Windows & Python version
1. Setting env variable in Windows PowerShell. Added:
#For Windows (PowerShell):
$env:MELTANO_ENVIRONMENT="dev"

2. Python version - 
I learned that Meltano must be pipx installed using a version of Python that is compatible with all of the extractors and loaders (see https://meltano.slack.com/x-p526315946325-3801798270198-3808513868754/archives/C01TCRBBJD7/p1657981158809849). I've added instructions for performing a pipx install with a specified version of Python, but I think this issue could use a bit more documentation to help the reader understand when they may need to use that command, especially because line 24 indicates everything _should_ work with Python 3.10, but tap-rest-api-msdk requires Meltano to be installed with 3.9.7 or earlier.